+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improved build process. (DanielRuf)
 - Added link to [demo](https://heiseonline.github.io/embetty/) to readme. (schliflo)
 - Added testing for latest node.js version. (DanielRuf)
+- Improved Youtube integration to use [privacy-enhanced mode](https://support.google.com/youtube/answer/171780?hl=en).
 - Upgraded dependencies.
 
 ## v1.0.5, 2018-05-28